Make getter methods more consistent
There were 2 kinds of getter methods in CloudControllerClientImpl based
on their behaviour when the entity they were supposed to return did not
exist - ones that threw exceptions (either IllegalArgumentExceptions or
CloudFoundryExceptions) and others that simply returned 'null'.

This commit fixes that inconsistency and now all of these methods throw
CloudFoundryExceptions in such cases.