Add additional getter methods
Recently, some getter methods were refactored to throw an exception
(instead of just returning 'null') if the entity they were supposed
to retrieve didn't exist.

This commit introduces new getter methods with a boolean flag, which
allows users to specify whether an exception should be thrown. These
should make the migration to the new version of the client easier.