Fix NPE for missing app binding  bindingForapplication.getBindingParameters()
This is a corner case where there are multiple parallel operations for app bindings.
Initially the app has binding to a service and while trying to compare existing and target
binding parameters, the app is unbind from the service.
The solution is to bind again the app to be sure that target binding parameters are applied.

Jira: LMCROSSITXSADEPLOY-1603