Request Space services instead of all services
Change the ListServicesRequest with ListSpaceServicesRequest in order to
fetch services, which are offered in the current working space. This
will prevent from trying to create service, which is not offered in the
current space, which results in 403. JIRA: LMCROSSITXSADEPLOY-1516

(cherry picked from commit 691b257893f886f60bb14489ff4a026d60475a10)